Fixes for linux build
- Fix paths when building docker image
- Swtich to cuda 11 from 12 for bitsandbytes
- Stop running install on all pip packages when they're already
  installed
- Misc fixes for build

+++ package/install_python_packages.sh
@@ -7,30 +7,39 @@ if python3 -c "import torch" 2>/dev/null; then
   echo "Torch is already installed."
 else
   echo "Torch is not installed. Installing torch, torchvision, and torchaudio..."
-  pip install -v --cache-dir $PIP_CACHE_DIR --prefix $PYTHONUSERBASE torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
+  pip install -v --cache-dir $PIP_CACHE_DIR --prefix $PYTHONUSERBASE torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
 fi
 
-# Check if AIRUNNER_ENABLE_OPEN_VOICE==1
+# # Check if AIRUNNER_ENABLE_OPEN_VOICE==1
 if [ "$AIRUNNER_ENABLE_OPEN_VOICE" == "1" ]; then
-  echo "Installing OpenVoice and MeloTTS..."
-  # Check if OpenVoice exists
-  if [ ! -d "OpenVoice" ]; then
-    git clone https://github.com/myshell-ai/OpenVoice.git
+  if python3 -c "import openvoice" 2>/dev/null; then
+    echo "OpenVoice is already installed."
+  else
+    echo "OpenVoice is not installed. Installing OpenVoice..."
+    # Check if OpenVoice exists
+    if [ ! -d "OpenVoice" ]; then
+      git clone https://github.com/myshell-ai/OpenVoice.git
+    fi
+    cd OpenVoice
+    pip install -v --cache-dir $PIP_CACHE_DIR --prefix $PYTHONUSERBASE .
+    cd ..
+    rm -rf OpenVoice
   fi
-  cd OpenVoice
-  pip install -v --cache-dir $PIP_CACHE_DIR --prefix $PYTHONUSERBASE .
-  cd ..
-  rm -rf OpenVoice
 
   # Check if MeloTTS exists
-  if [ ! -d "MeloTTS" ]; then
-    git clone https://github.com/myshell-ai/MeloTTS.git
+  if python3 -c "import melo" 2>/dev/null; then
+    echo "MeloTTS is already installed."
+  else
+    echo "MeloTTS is not installed. Installing MeloTTS..."
+    if [ ! -d "MeloTTS" ]; then
+      git clone https://github.com/myshell-ai/MeloTTS.git
+    fi
+    cd MeloTTS
+    git checkout v0.1.2
+    pip install -v --cache-dir $PIP_CACHE_DIR --prefix $PYTHONUSERBASE .
+    cd ..
+    rm -rf MeloTTS
   fi
-  cd MeloTTS
-  git checkout v0.1.2
-  pip install -v --cache-dir $PIP_CACHE_DIR --prefix $PYTHONUSERBASE .
-  cd ..
-  rm -rf MeloTTS
 
   echo "Downloading unidic..."
   python3 -m unidic download
@@ -42,13 +51,13 @@ else
   pip uninstall melotts -y
 fi
 
-# Install python packages at runtime
-
 # Check if airunner is installed and which version, if not installed or outdated, install the latest version
 if python3 -c "import airunner" 2>/dev/null; then
   echo "Airunner is already installed."
-  INSTALLED_VERSION=$(python3 -c "import airunner; print(airunner.__version__)")
-  REQUIRED_VERSION=$(grep 'version=' /app/setup.py | cut -d "'" -f 2)
+  INSTALLED_VERSION=$(pip show airunner | grep Version | cut -d ' ' -f 2)
+  REQUIRED_VERSION=$(grep 'version=' /app/setup.py | sed -n "s/.*version=['\"]\([^'\"]*\)['\"].*/\1/p")
+  echo "Installed version: $INSTALLED_VERSION"
+  echo "Required version: $REQUIRED_VERSION"
   if [ "$INSTALLED_VERSION" != "$REQUIRED_VERSION" ]; then
     echo "Airunner version $INSTALLED_VERSION is installed, but version $REQUIRED_VERSION is required."
     pip install -v --cache-dir $PIP_CACHE_DIR --prefix $PYTHONUSERBASE -e .[all_dev] \
@@ -59,11 +68,23 @@ if python3 -c "import airunner" 2>/dev/null; then
     echo "Airunner version $INSTALLED_VERSION is already installed."
   fi
 else
+  echo "Airunner is not installed. Installing the latest version..."
   pip install -v --cache-dir $PIP_CACHE_DIR --prefix $PYTHONUSERBASE -e .[all_dev] \
       -U langchain-community \
       -U mediapipe
   pip install -v --cache-dir $PIP_CACHE_DIR --prefix $PYTHONUSERBASE -U timm
 fi
 
-python3 -c "import nltk; nltk.download('punkt')"
-python3 -c "from accelerate.utils import write_basic_config; write_basic_config(mixed_precision='fp16')"
+if [ -d "/home/appuser/nltk_data/tokenizers/punkt" ]; then
+  echo "NLTK punkt tokenizer is already installed."
+else
+  echo "NLTK punkt tokenizer is not installed. Installing..."
+  python3 -c "import nltk; nltk.download('punkt')"
+fi
+
+if [ -f "/home/appuser/.local/share/airunner/.cache/huggingface/accelerate/default_config.yaml" ]; then
+  echo "Accelerate config file already exists."
+else
+  echo "Accelerate config file does not exist. Creating..."
+  python3 -c "from accelerate.utils import write_basic_config; write_basic_config(mixed_precision='fp16')"
+fi
